Implemented MPU6500 and AK8975(AK8963) libraries.
Those libraries for IMU and MAG are replaced by MPU6500 and AK8975(AK8963). Changing those data format for IMU and MAG from int to float.

+++ software/arduino-platformio-vscode/lib/App/Pages/SystemInfos/SystemInfosModel.cpp
@@ -69,9 +69,9 @@ void SystemInfosModel::GetGPSInfo(
 
 void SystemInfosModel::GetMAGInfo(
     float* dir,
-    int* x,
-    int* y,
-    int* z
+    float* x,
+    float* y,
+    float* z
 )
 {
     HAL::MAG_Info_t mag = { 0 };
@@ -96,7 +96,7 @@ void SystemInfosModel::GetIMUInfo(
     snprintf(
         info,
         len,
-        "%d\n%d\n%d\n%d\n%d\n%d",
+        "%0.2f\n%0.2f\n%0.2f\n%0.2f\n%0.2f\n%0.2f",
         imu.ax,
         imu.ay,
         imu.az,
